+++ service/src/main/java/com/bestomb/common/response/goods/GoodsBo.java
@@ -8,6 +8,8 @@ import com.bestomb.common.util.CalendarUtil;
 import com.bestomb.entity.GoodsWithBLOBs;
 import com.bestomb.service.IDictService;
 
+import java.math.BigDecimal;
+
 public class GoodsBo {
 	
 	private Logger logger = Logger.getLogger(this.getClass());
@@ -21,7 +23,7 @@ public class GoodsBo {
 	
 	private String id; // 商品ID
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String belongs; // 商品所属 （1：会员商品、2：系统商品）
     private String type; // 商品类型 （1：陵园装饰主题、2：陵园、3：陵园存储、4：交易币、5：动物、6：植物）
     private String modelId; // 商品模型ID
@@ -75,12 +77,15 @@ public class GoodsBo {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Double getPrice() {
+
+	public BigDecimal getPrice() {
 		return price;
 	}
-	public void setPrice(Double price) {
+
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
+
 	public String getBelongs() {
 		return belongs;
 	}
